requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
blinker==1.4
certifi==2022.12.7
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==3.4.8
dash==2.7.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
distlib==0.3.6
distro==1.7.0
Django==3.2.16
#django-plotly-dash==2.1.3
dpd-components==0.1.0
filelock==3.9.0
Flask==2.2.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.1
more-itertools==8.10.0
msgpack==1.0.4
netifaces==0.11.0
numpy==1.24.0
oauthlib==3.2.0
pandas==1.5.2
phonenumbers==8.13.3
platformdirs==2.6.2
plotly==5.11.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytube==12.1.2
pytz==2022.7
PyYAML==5.4.1
redis==4.4.0
requests==2.28.1
SecretStorage==3.3.1
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
systemd-python==234
tenacity==8.1.0
ufw==0.36.1
urllib3==1.26.13
virtualenv==20.17.1
wadllib==1.3.6
Werkzeug==2.2.2
zipp==1.0.0
